Some more code re-arrangements
